+++ internal/bundler/linker.go
@@ -5322,13 +5322,14 @@ func (c *linkerContext) generateChunkCSS(chunks []chunkInfo, chunkIndex int, chu
 			}
 
 			cssOptions := css_printer.Options{
-				MinifyWhitespace:  c.options.MinifyWhitespace,
-				ASCIIOnly:         c.options.ASCIIOnly,
-				LegalComments:     c.options.LegalComments,
-				SourceMap:         c.options.SourceMap,
-				AddSourceMappings: addSourceMappings,
-				InputSourceMap:    inputSourceMap,
-				LineOffsetTables:  lineOffsetTables,
+				MinifyWhitespace:    c.options.MinifyWhitespace,
+				ASCIIOnly:           c.options.ASCIIOnly,
+				LegalComments:       c.options.LegalComments,
+				SourceMap:           c.options.SourceMap,
+				UnsupportedFeatures: c.options.UnsupportedCSSFeatures,
+				AddSourceMappings:   addSourceMappings,
+				InputSourceMap:      inputSourceMap,
+				LineOffsetTables:    lineOffsetTables,
 			}
 			compileResult.PrintResult = css_printer.Print(asts[i], cssOptions)
 			compileResult.sourceIndex = sourceIndex